Fixes type conversion to still compile with java 1.8
Converting from one type to an incompatible one is no longer allowed in
Java 1.8. Here, a list of FormalConcept (an interface) instances are not
known to the compiler to be instances of Quality at runtime, and so the
conversion is disallowed.

The fix here is simple - convert to a non-generic list type first, and then
back to the desired generics template type.